Extract address autofill logic to base module
